requirements.txt
anyio==3.7.1
certifi==2023.5.7
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ojson==0.1.0
Pillow==10.0.0
sniffio==1.3.0
toml==0.10.2
websockets==11.0.3
wget==3.2
